pyproject.toml
[project]
name = "travel-mcp-server"
version = "0.1.0"
description = "MCP server for travel aggregation - flights, hotels, and more"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Travel MCP Contributors" }
]
dependencies = [
    "mcp>=1.0.0",
    "amadeus>=8.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/travel_mcp"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
